// src/app/shared/swagger/model/Content.ts
/**
 * Yaturu
 * ## Welcome  This is a place to put general notes and extra information,
 *  for internal use.  To get started designing/documenting this API,
 *  select a version on the left. ## Welcome  This is a place to put general notes and extra information,
 *  for internal use.  To get started designing/documenting this API, select a version on the left.
 *
 * OpenAPI spec version: 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import * as models from './models';

export interface Content {
    name?: string;

    url?: string;

    thumbnailUrl?: string;

    trackNumber?: number;

    contentType?: Content.ContentTypeEnum;

    description?: string;

    question?: models.Question;

    languages?: Array<string>;

    duration?: number;

    key?: string;

}
export namespace Content {
    export enum ContentTypeEnum {
        Image = <any>'image',
        Audio = <any>'audio',
        Video = <any>'video',
        Text = <any>'text',
        Question = <any>'question',
        Cupon = <any>'cupon'
    }
}